fix: remove tmp file before quit

+++ packages/webpage/common/electron.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
-
+// for type hints
 const _electron = {
   mkDir(path: string) {
     fs.mkdirSync(path)
@@ -15,6 +15,7 @@ const _electron = {
     fs.writeFileSync(path, data)
   },
   remove(path: string) {
+    if (!fs.existsSync(path)) return
     if (fs.statSync(path).isFile()) fs.unlinkSync(path)
     else {
       const files = fs.readdirSync(path)
@@ -30,12 +31,21 @@ const _electron = {
     port: number,
     hostname: string
   ): Promise<{ type: 'local' | 'network'; address: string; port: number }[]> {
-    return new Promise(resolve => {
+    return new Promise((resolve, reject) => {
       const middleware = connect()
       _server = http.createServer(middleware)
       middleware.use(sirv(path))
       middleware.use(sirv(join(__dirname, './page')))
-
+      const onError = (e: Error & { code?: string }) => {
+        if (e.code === 'EADDRINUSE') {
+          _server!.removeListener('error', onError)
+          reject(new Error(`Port ${port} is already in use`))
+        } else {
+          _server!.removeListener('error', onError)
+          reject(e)
+        }
+      }
+      _server.addListener('error', onError)
       _server.listen(port, hostname, () => {
         const address = Object.values(os.networkInterfaces())
           .flatMap(nInterface => nInterface ?? [])
@@ -50,6 +60,7 @@ const _electron = {
               : 'network'
             return { type, address: detail.address, port }
           })
+        _server!.removeListener('error', onError)
         return resolve(address)
       })
     })
@@ -64,6 +75,9 @@ const _electron = {
     return new Promise(resolve =>
       ipcRenderer.on('openDialogReply', (_, args) => resolve(args))
     )
+  },
+  sendTmpFile(path: string) {
+    ipcRenderer.send('tmpFile', path)
   }
 }
 // @ts-check
